+++ src/engines/event-contract-collection-runner-rehearsal-operation-control/index.ts
@@ -1,3 +1,4 @@
 export * from "./EventContractCollectionRunnerRehearsalOperationControlEngine";
 export * from "./EventContractCollectionRunnerRehearsalOperationControlConsole";
 export * from "./EventContractCollectionRunnerRehearsalOperationPreflight";
+export * from "./EventContractCollectionRunnerRehearsalOperationVerification";